backend/requirements.txt
fastapi
uvicorn
pathway
google-generativeai
python-multipart
